task1/django_backend/taskapp/requirements.txt
Django==5.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.2
Pillow==10.0.1  
requests==2.31.0  
